aacenc: WIP support for PCEs
Hopefully whoever wants to have support for crazy formats can help.
The table in aacenc.h (temporary position) tells the encoder what
to put in the bitstream and how to encode. Problem is, the specifications
dont specify anything. Thats because I've not been able to find any bloody
specifications and had to work with what the decoder does. And there was
plenty of guessing there because the decoder does some magic on layout_map
which I can't even figure out nor even know if its correct (it seems to be
for the formats I've tested).
Then there's the problem with the exact order that the channels have to be in.
Again a guessing game, since you essentially have no idea what the index part
of the map is supposed to be, whether it has to be incremented starting from
the first channel or reset upon every front/side/back channel groups. At
least the map to instruct the encoder's straightforward.

Anyway, help appreciated.

Applies cleanly on 543142990b6f7b8757753c13ea6dbc56275c5c7e, but should
work fine with newer versions.